地址、商品

+++ src/main/java/com/graduation/panda/model/UserAddress.java
@@ -14,7 +14,7 @@ public class UserAddress {
     private boolean dataFlag;
     private Date createTime;
 
-    public long getAddressId() {
+    public int getAddressId() {
         return addressId;
     }
